Create Binary Tree to CDLL (Medium).cpp
Given a binary tree (BT), the task is to convert this to a Circular Doubly Linked List (CDLL) in-place. The left and right pointers in nodes will be used as previous and next pointers respectively in CDLL. The order of nodes in CDLL must be the same as Inorder of the given Binary Tree. The first node of Inorder traversal (leftmost node in BT) must be the head node of the CDLL.

Examples:

Input:
      1
    /   \
   3     2
Output:
3 1 2 
2 1 3

Explanation: After converting tree to CDLL when CDLL is is traversed from head to tail and then tail to head, elements are displayed as in the output.
Input:
     10
   /    \
  20    30
 /  \
40  60
Output:
40 20 60 10 30 
30 10 60 20 40

Explanation: After converting tree to CDLL, when CDLL is is traversed from head totail and then tail to head, elements are displayed as in the output.