Create Linked List to Binary Tree (Medium).cpp
Given a Linked List Representation of Complete Binary Tree. Your task is to construct the Binary tree from the given LinkedList and return the root of the tree.
The result will be judged by printing the level order traversal of the Binary tree.

Note: The complete binary tree is represented as a linked list in a way where if the root node is stored at position i, its left, and right children are stored at position 2*i+1, and 2*i+2 respectively. H is the height of the tree and this space is used implicitly for the recursion stack.

Examples:

Input: linkedList: 1 -> 2 -> 3 -> 4 -> 5
 
Output: root[] = [1, 2, 3, 4, 5]
Explanation: The tree would look like 
 
Now, the level order traversal of the above tree is 1 2 3 4 5.
Input: linkedList: 5 -> 4 -> 3 -> 2 -> 1

Output: root[] = [5, 4, 3, 2, 1]
Explanation: The tree would look like
 
Now, the level order traversal of the above tree is 5 4 3 2 1.
Input: linkedList: 10 -> 20
Output: root[] = [10, 20]
Explanation: The Binary Tree formed is:
     10
    /
   20
The level order traversal of this tree is [10, 20].
Constraints:
1 <= numbers of nodes <= 106
1 <= node->data <= 104Given a Linked List Representation of Complete Binary Tree. Your task is to construct the Binary tree from the given LinkedList and return the root of the tree.
The result will be judged by printing the level order traversal of the Binary tree.

Note: The complete binary tree is represented as a linked list in a way where if the root node is stored at position i, its left, and right children are stored at position 2*i+1, and 2*i+2 respectively. H is the height of the tree and this space is used implicitly for the recursion stack.

Examples:

Input: linkedList: 1 -> 2 -> 3 -> 4 -> 5
 
Output: root[] = [1, 2, 3, 4, 5]
Explanation: The tree would look like 
 
Now, the level order traversal of the above tree is 1 2 3 4 5.
Input: linkedList: 5 -> 4 -> 3 -> 2 -> 1

Output: root[] = [5, 4, 3, 2, 1]
Explanation: The tree would look like
 
Now, the level order traversal of the above tree is 5 4 3 2 1.
Input: linkedList: 10 -> 20
Output: root[] = [10, 20]
Explanation: The Binary Tree formed is:
     10
    /
   20
The level order traversal of this tree is [10, 20].
Constraints:
1 <= numbers of nodes <= 106
1 <= node->data <= 104